@@ -1,14 +1,16 @@
-import React from 'react';
-import {BrowserRouter as Router,Route} from 'react-router-dom';
+import React from "react";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
 
-import Join from './components/Join/Join';
-import Chat from './components/Chat/Chat';
+import Join from "./components/Join/Join";
+import Chat from "./components/Chat/Chat";
 
-const App=()=>(
-    <Router>
-        <Route path="/" exact component={Join}/>
-        <Route path="/chat" exact component={Chat}/>
-    </Router>
+const App = () => (
+  <BrowserRouter>
+    <Routes>
+      <Route path="/" element={<Join />} />
+      <Route path="/chat" element={<Chat />} />
+    </Routes>
+  </BrowserRouter>
 );
 
-export default  App;
+export default App;
